Fix chapter passthru problems with some files. If the video stream is not the first track in the file, chapters were lost.
During scan, we identify which track is video and stash this in title.
While reading, when a chapter is found we want to tag the next video buffer.
But the video track id stored in the title was not being applied when
opening the file for reading, so the chapter mark always went on track id 0.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@3889 b64f7644-9d1e-0410-96f1-a4d463321fa5