fix a crash in HandBrakeCLI
When the audio tracks are omitted (no -a), the cli uses the first source
audio track and replicates it for each audio output codec.  The way it
knows what to replicate is to look at audio->in.track for the last audio
that was added to the job.  This track number was being copied from the
title's in.track. But title track numbers are not guaranteed to be
sequential (e.g. scan drops tracks it can't decode).  So when adding an
audio to a job, set in.track to the value of in.track that is passed in
the audio_config structure passed to hb_audio_add.

git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4260 b64f7644-9d1e-0410-96f1-a4d463321fa5