Fix fontconfig search path for linux and osx
It was searching for .../contrib/etc/fonts.  Needed to be set to the
standard search paths on the respective systems.

On windows, it already searches the directory fonts under the
executables directory.  We need to put a copy of fonts.conf there. This
can be copied from contrib/etc/fonts/fonts.conf

git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@4405 b64f7644-9d1e-0410-96f1-a4d463321fa5