libav: solaris build fixes
1. Link in /usr/lib/values-xpg6.o to get proper POSIX.1-2001 behaviour
from libc -- this is controversial within the GCC camp as it "breaks old code"
that doesn't assume it has been added. Some libc functions behave very
differently with and without this linked. Since we're linking statically to
the library, this is not a significant concern.

2. MPC8 disable constraint is clamped down to only apply on SPARC.

thanks to icchan for patch.
https://reviews.handbrake.fr/r/367/


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@5138 b64f7644-9d1e-0410-96f1-a4d463321fa5