libhb: Fix issue with silence insertion when using FLAC encoder
When we insert silence to maintain sync, we typically need to insert about
70ms.  But the default flac frame duration is very large (>=96ms).  This is
causing us to overshoot the target and we immediately drop a frame.  Dropping
a frame undershoots and we insert a frame... ad infinitum.

This patch changes the behavior of silence insertion for non-passthru audio.
Instead of inserting full frames of silence, the last "frame" inserted may be
a partial.  This lets us insert a finer grained silence duration and eliminates
the problem with large flac frames.


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@5188 b64f7644-9d1e-0410-96f1-a4d463321fa5