MacGui: forbid including passthru PGS subtitle streams in MP4.
When adding a PGS track and the container is MP4, the track is automatically set to burned-in.

When adding an additional PGS track and the container is MP4, the track is removed and an alert dialog is displayed.

When switching from MKV to MP4, the first PGS track is converted to burned-in (any non-PGS track that was set to burned-in is reset to passthru) and additional tracks are removed; an alert dialog is also displayed.



git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@5325 b64f7644-9d1e-0410-96f1-a4d463321fa5