WinGui: Picture Settings Tab and Preset Behaviours changed to be consistent with the macgui.
- None:  
Nothing will happen to the W/H/KeepAR when changing preset.
Changing Title or source will change the Width to the source width and turn Keep AR on.

- Custom
Changing preset will change the resolution to be no more than what the Max W and H were set to. Keep AR flag will be honour'd, so if it's turned off, the output AR could be incorrect.
Set the W/H to a large number for the equiv of No Limit.

- Source Max
Same idea as custom, but limited to the resolution of the current source instead of a custom W/H


This now means that importing a mac or linux preset provides consistent behaviour.




git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk/win/CS@5525 b64f7644-9d1e-0410-96f1-a4d463321fa5