libhb: improve video and audio encoder names.
When we only have one encoder for a particular codec:

- don't add (library_name) in the name
- don't prefix with ff or otherwise use the library name in the short name

AAC (ffmpeg) -> AAC (avcodec): works for both Libav and FFmpeg.
consistency: ffaac -> av_aac (like ca_aac, fdk_aac and so on)

VP3 (Theora) -> Theora: Theora is based on VP3, but is not exactly VP3.

manicure updated to handle the new names, but built-in presets left untouched (this will require WinGUI updates).



git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@5586 b64f7644-9d1e-0410-96f1-a4d463321fa5