libhb: fix scanning some dvds e.g. star trek into darkness
This reverts part of a patch that was originally written for dvdread
access.  So we need to watch for regressions.

The patch was originally written for dvdread access and my not actually be
a bug in dvdnav.  So I have reverted it in dvdnav only. I also tried to 
reproduce the original bug in dvdread, but was unsuccessful. So I'm
wondering if the original bug also got "fixed" by subsequent changes.

The original bug report is here:
https://forum.handbrake.fr/viewtopic.php?f=4&t=9758&start=0

The original fix is here:
https://trac.handbrake.fr/changeset/2256


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@5787 b64f7644-9d1e-0410-96f1-a4d463321fa5