LinGui: rework title list
Title list is now linked to a settings list.  There is one settings dict
per title.  The settings list is populated when scan completes.  As you
change titles, the settings for that title populate the UI.  Changes
made are saved in the settings and restored if you go to another title
and then return.

This is preparation for improvements to "Add All to Queue" feature.
Settings for each title can be customized before adding to the queue.
All titles don't have to use the same preset or settings.

git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6046 b64f7644-9d1e-0410-96f1-a4d463321fa5