libhb: repair split packets in reader
We split PES packets when there is a PCR change in the middle of the
packet.  This works fine for audio and video where the decoder parses
the ES to find frame boundaries.  But it does not work for some decoders
such as PGS subtitles.  So mark split buffers and reassemble them in
reader after processing the PCR change.