sync: gut and rewrite
sync.c was difficult to read, fragile, and prone to difficult to
diagnose bugs (see
https://forum.handbrake.fr/viewtopic.php?f=12&t=33147)

This rewrite simplifies the code, removes signals, locking and yield
that probably cause the above problem and is much more flexible.  It
fixes a wider variety of timestamp issues than before and is much easier
to extend if other timestamp analysis is desired.