deinterlace: add conditional deinterlacing of interlaced frames
Only deinterlace frames that are marked as interlaced when mode
HB_YADIF_AUTO is set in the filter settings.

Currently, this would cause no frames to be deinterlaced since all
frames are marked non-interlaced.  This option will become useful in a
subsequent patch that splits comb detection into it's own filter.