logging: rename hb_error to hb_spam_log and create new hb_error
hb_error was originally written to consolidate spammy log messages by
checking for duplicates and only showing first occurance plus last
occurance and count.  It also had a callback interface that the
frontend could use to capture these messages.  But the spam isn't really
useful to the frontend. So I've simplified it to only do message
deduplication.

The new hb_error is connected to the callback and is used more frugally
so should be more appropriate for frontend error feedback.