sync: speed up p-to-p finalization when there are subtitles
When doing point-to-point encoding, subtitles can cause a long
delay in finishing the job when the stop point is reached.  This
is due to the sparse nature of subtitles.  We may not even see
any additional subtitle till we reach the end of the file.

So when all audio and video streams have reached the end point,
force the termination of all subtitle streams by pushing an
end-of-stream buffer into each subtitles input fifo.
This will cause each subtitle sync worker to wake and return
HB_WORK_DONE.