vfr: add cfr frame drop debugging logs
These logs can be enabled by uncommenting HB_DEBUG_CFR_DROPS at the top
of the file.

If you have any sources that use progressive frame upsampling that you
would like to test with this new frame drop algo, enable these debug
logs to get full details of what frames are dropped and passed.

They show which frames are dropped and the metrics that the decision to
drop is based on.  They also show which frames are passed and the
cadence of passed vs. dropped frames.