sync: don't delay subtitles when duration is known
Delayed subtitles were causing incorrect muxing in mkv.  The mkv muxer
writes chunks where all samples should be relative to a chunk's base
timestamp.  When the subtitle is delayed long enough for a new chunk to
start before it gets muxed, the calculated offset to the chunk's base
time is negative (which is illegal).

Note that this is still a possibility with subtitles that must be
delayed (e.g. CC and VOBSUB) because the duration is not known until
the next subtitle's start time is known. The only fix for this would be
to add a special subtitle parsing pass that caches subtitle timestamps
before the main encoding pass is performed.