libav: explicitly write encoder delay and preroll to mp4
This will make it possible to read encoder delay back and drop the
samples appropriately.

Writing preroll sample group to the mp4 fixes post-processing of the
file with Apple tools.  If the roll sample group is not present, Apple
tools will apply an implicit rule to remove encoder delay which results
in the delay being dropped twice.