sync: fix small video dejitter error
This very small error snowballs into a crash in x264 :-p

If the amount of jitter on the first frame in the queue was small
(about 1 tick) then jitter would not be removed from that frame.
This extra tick of jitter can appear on different frames depending
on when frame arrives and how much has been queued.  This very small
amount of randomness lead to problems in the VFR filter.  A frame
duration difference as small as 1 tick can lead to an extra frame
getting duplicated when doing CFR.  When doing 2 pass encoding, this
extra frame causes x264 to crash at the end of the 2nd pass.

(cherry picked from commit 31d0389758bd506a6c349d85c0d3191880d0dea4)