scan: fix very slow scanning for some files
The threshold in bytes for when to give up trying to decode a frame was
too big for a lot of streams.  It was made large to accomodate 4K raw
video.  Instead of counting bytes, count frames fed to the decoder.
This is more consistant regardless of video resolution and codec.