reader: fix live preview generation for some streams
Fixes https://github.com/HandBrake/HandBrake/issues/685

For some stream types, libav does not seek all streams within the file
to the same position.  So we get data from streams that is prior to the
desired preview start position leaking through. The result, for example,
is a preview that has audio that starts from the beginning of the file
and video that starts much later at the desired start pos.