LinGui: Add "Category" dropdown to preset save dialog
This replaces the "New Folder" option in the presets menu.  It enforces
the folder structure we have agreed to and hopefully helps the user keep
things organized.

Note that users are allowed to save a custom preset to the same
"Category" as an official preset.  When they do this, a new custom
category is created with the same name and the preset is saved in that
folder.