Performed some restructuring of C#/.NET folders.
Moved win\CS to win\, as there is only ever going to be one implementation on an OS, and other languages should be added in the same folder / VS sln.

Moved HandBrake.ApplicationServices into netstandard/HandBrake.ApplicationServices, this is because it is no longer limited to being for Windows only. Added it's own .sln, for if you want to compile it separately from the Windows Solution.