encx264: fix encoding of very short duration frames
We had a hash of limited size where we stored frame durations in
encx264.  This hash has not been necessary since we moved to using
libavformat for muxing.  Since the limited size of the hash put
constraints on how short frame durations could be, I've eliminated it.
This means that after the encoder, hb_buffer_t s.stop and s.duration are
no longer valid.  But since they are not used during muxing, this is not
a problem.

Fixes https://github.com/HandBrake/HandBrake/issues/1374