LinGui: add search function to language popups
The language list is long and displays native names while sorted by
English names, so it is difficult to find a language.  This adds the
ability to "search" for a language based on it's initial character.
Typing a character will jump to the next entry that starts with that
character in its native name.