LinGui: change language lists from GtkListBox to GtkTreeView
The GtkTreeView is more flexible and more capable.  Languages can be
searched for in the GtkTreeView by focusing the widget and typing what
you want to find.  Double clicking an item in the list will add or
remove it from the "selected" languages list in audio and subtitle track
selection settings.