ffmpeg: fix currupt h.264 preview frames
The ffmpeg developers added a "heuristic" to the h.264 recovery point
code I wrote to prevent display of corrupt frames.  The heuristic
doesn't work very reliably, so I've disabled it.

When I get a chance I'll see if I can come up with a better way to
handle what the heuristic was added for.