bd: handle "broken units" better
libbluray detects and skips broken units (6114 byte chunks).  But it
returns 0 in such cases.  We were interpreting this as the end of the
title and signalling a successful encode.

Instead, look for BD_EVENT_END_OF_TITLE for end of title detection.
And perform read retries when bd_read returns 0.