fix race in getting sequence_id of completed job
The sequence_id was only available for the WORKING state and not the
WORKDONE state.  But frontends poll for status periodically and can miss
all status updates for the WORKING state if the file is very short or an
error occurs early during transcoding.  When WORKING status is missed,
there was no way to know the sequence_id associated with the WORKDONE
status.