flatpak: fix access to gvfs mounted filesystems
An additional access permission is needed to access the gvfs mount
point.

For the record, the "Right" (TM) way to do this is to use
GtkFileChooserNative dialog which will automatically use the flatpak
file chooser portal to access files.  But this portal currently is
missing features that would cripple HandBrake.  These features are:

1. Folder selection.  We require this for destination selection and for
batch processing.
2. Set current highlighted file in file chooser browser.  Without this,
a user processing a series of videos that are all in the same foder has
to navigate from the default "Recent files" selection to the folder
where their files are for each file.

Also note that GtkFileChooserNative transparently falls back to a
regular file chooser when the action requested is not supported.  So
sometimes the user would have access to remote filesystems, and
sometimes not with no obvious reason for the difference.  Not a good
user experience.

(cherry picked from commit d4761a9f2f07fe02935089356fd847530a8eec57)