LinGui: work around GTK slider value formatting bug
When the range of a slider changes, GTK used to sample all the
possible values to determine the correct amount of screen space to
allocate for the value strings. Some *genius* decided it would be
more effecient to just sample the first and last value which means
that if certain characters are wider than others and the middle
values happen to use those characters, the space allocated is
too small and the string wraps to the next line or is truncated.

So, we have to randomly add some extra space to the first and
last value string in order for the string to be displayed properly.
WTF guys!