vfr: improve duplicate frame drop detection
More strictly limit the range of frames inspected when choosing a frame
to drop.

We were getting too many false positives which resulted in dropping
good frames. Limiting the range results in emitting good frames
before they can become the worst of a collection of other good
frames, and therefore be evaluated as droppable.