LinGui: gtk4 port, fix summary tab preview image scaling
GTK4 added a new widgit GtkPicture that is almost identical to GtkImage.
The main difference appears to be that GtkImage now scales down the size
of the given pixbuf using some undocumented algorithm.  It's filling
some of the space available, but not all of it.  But it isn't a hard
size limit because making the window bigger continues to make the image
bigger.

GtkPicture shows the pixbuf unscaled as expected. So use GtkPicture for
the preview image :*(