ffmpeg: make track names visible regarless of locale
Apple will not show a track name if it doesn't match the current locale
or it's language is not "und".  This patch results in the language being
set to "und".

writes 3gpp track names for all language variations in metadata.
when the language is not specified, uses "und" as default.

HandBrake doesn't actually allow multiple track name languages, but one
of the side effects of this patch is to use "und" when the track name
language is unspecified. Before this patch, language specific variants
of metadata were ignored and "eng" was always set as the language.