qsv: Enforce GOP parameter constraints for HEVC
The hierarchical B pyramid used in HEVC requires the GOP size be a power of 2 and at most 256. Furthermore, the number of reference frames for the B pyramid is defined in the unsupported CodingOption3 so must be fixed at the max defined for HEVC of 15. If not using a b pyramid these constraints do not apply and the hw default can now be obtained by providing 0s or leaving them unset.

CQP defaults to b pyramid if using HEVC wheras ICQ defaults to it's simple frequent GPB frame detection. No constraints are applied to ICQ .. defaults are ideal.