qsv: Detect supported params for hw codecs
All encoders take all structures, they just tend to use radically different parameters most of which get reset on the call to init if they are supplied invalid. This change updates the detection code to use this reset to detect support and allows HEVC to use the CodingOption2 structure to configure the B pyramid. This fix is minimal and is only applied to parameters needed for HEVC.