preset: Use x265 aq-mode 1 for official presets.
We have received numerous reports about x265 in HandBrake 1.3.0 creating ~30% larger files with no perceived quality gain, and have confirmed this to be true in most cases. This appears to be the result of x265 recently switching to use aq-mode 2 by default. This change makes the behavior of the official presets more closely match that of HandBrake 1.2.2 and earlier. This seems to be the best course of action both technically and for the user experience until the default aq-mode improves.

(cherry picked from commit e6d29a01a35ff3c942b4a1e5693bf6b9e678cc37)