cropscale: fix color range conversion
range was already getting converted to limited in decavcodec.c

The range of each frame is provided in AVFrame.color_range and we are
already propagating that through the pipeline.  Manually setting
"in_range" in the "scale" filter was overriding that per frame setting.

Fixes https://github.com/HandBrake/HandBrake/issues/2561