Enbale potential for conservative cropping per #2219
Include a flag that if set, overrides the autocrop value to prefer the most conservative cropping possible. Also reduce ambiguity in the comment by changing "loose" cropping to "conservative" cropping (and "strict" to "aggressive").

This is a minimal change and would still require a GUI update to become a live option.