LinGUI: Remove update checker
The update checker didn't work and could not be built on any recent
Linux distribution. Linux updates are handled by Flatpak or the
distribution.
configure.py will error out if the --disable-gtk-update-checks
flag is used. This is to avoid the flag being passed to make and
giving an error anyway.