libdovi: permit building with pre-downloaded deps
Dependencies can be pre-downloaded and bundled into the tar using the
'cargo vendor' command. In order to use these pre-downloaded crates, the
outpout of 'cargo ventor' must be sent to the file '.cargo/config.toml'
in the directory to be built. But, cargo will only read this file if the
current working directory when cargo is executed contains the .cargo
directory. So make sure this is the case by changing to the directory
to be built instead of using the --manifext-path option.