Cli: correctly set up preset values
When there are more settings specified than items in the AudioList list, HandBrake needs
to add more items. In the case where the AudioList in the preset has items, we just
duplicate the last item and leave it at that. In the case there the AudioList is empty, we
want to use the last setting supplied on the command line in all the remaining items.