gtk: Fix preview black borders on Wayland
When GTK is running on GNOME Wayland, windows with standard
title bars use client-side decorations. This means the height of
the title bar is taken into account when requesting a resize.
Instead of attempting to calculate the title bar height, we just
set the width and let the fixed aspect ratio of the video frame
take care of the rest.

Fixes #6090.